# src/api/main.py
"""
Main API implementation for Agent360.
Provides RESTful endpoints for agent interaction and management.
"""
from typing import Dict, Any, List
from fastapi import FastAPI, HTTPException, Depends
from fastapi.middleware.cors import CORSMiddleware
from fastapi.security import OAuth2PasswordBearer
from pydantic import BaseModel
import logging
import json
from datetime import datetime

# Initialize FastAPI app
app = FastAPI(
    title="Agent360 API",
    description="Enterprise-grade agent infrastructure API",
    version="1.0.0"
)

# Configure CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Configure authentication
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class AgentRequest(BaseModel):
    """Model for agent execution requests."""
    input: Dict[str, Any]
    tools: List[str] = []
    timeout: int = 300

class AgentResponse(BaseModel):
    """Model for agent execution responses."""
    output: Dict[str, Any]
    execution_time: float
    tool_usage: Dict[str, int]

@app.post("/api/v1/agent/execute", response_model=AgentResponse)
async def execute_agent(
    request: AgentRequest,
    token: str = Depends(oauth2_scheme)
):
    """
    Execute an agent task.
    
    Args:
        request: Agent execution request
        token: Authentication token
        
    Returns:
        Agent execution response
    """
    try:
        start_time = datetime.now()
        
        # TODO: Implement agent execution logic
        # This should:
        # 1. Validate the request
        # 2. Initialize required tools
        # 3. Execute the agent workflow
        # 4. Monitor and collect metrics
        # 5. Return results
        
        raise NotImplementedError("Agent execution not implemented")
        
    except Exception as e:
        logger.error(f"Agent execution failed: {str(e)}")
        raise HTTPException(
            status_code=500,
            detail=str(e)
        )

@app.get("/api/v1/agent/tools")
async def list_tools(token: str = Depends(oauth2_scheme)):
    """
    List available tools.
    
    Args:
        token: Authentication token
        
    Returns:
        List of available tools and their metadata
    """
    try:
        # TODO: Implement tool listing logic
        raise NotImplementedError("Tool listing not implemented")
        
    except Exception as e:
        logger.error(f"Failed to list tools: {str(e)}")
        raise HTTPException(
            status_code=500,
            detail=str(e)
        )

@app.get("/health")
async def health_check():
    """Health check endpoint."""
    return {"status": "healthy", "timestamp": datetime.now().isoformat()}